+++ src/protocols/dhcpv6/encode.c
@@ -171,7 +171,7 @@ static ssize_t encode_value(fr_dbuff_t *dbuff,
 			if ((da->flags.subtype == FLAG_ENCODE_PARTIAL_DNS_LABEL) &&
 			    (*(fr_dbuff_current(&p) + fr_dbuff_current(&p)[0] + 1) == 0)) {
 				fr_dbuff_set_to_start(&work_dbuff);
-				fr_dbuff_advance(&work_dbuff, slen - 1);
+				fr_dbuff_advance(&work_dbuff, (size_t)slen - 1);
 			}
 			break;
 		}
@@ -321,7 +321,7 @@ static ssize_t encode_value(fr_dbuff_t *dbuff,
 			fr_dict_attr_t **u;
 
 			memcpy(&u, &c, sizeof(c)); /* const issues */
-			memcpy(u, &vp->da, sizeof(vp->da));			
+			memcpy(u, &vp->da, sizeof(vp->da));
 		}
 		FALL_THROUGH;
